backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4flask==3.0.0
flask-cors==4.0.0